+++ CHANGELOG.md
@@ -12,8 +12,10 @@ and this project adheres to
 
 - `mpvd push -p`/`--playlist` flag to push tracks to an XSPF playlist file
   instead of the current playlist.
-- `mpvd remove -p`/`--playlist` flag to remove tracks from an XSPF playlist
-  file instead of the current playlist.
+- `mpvd remove -p`/`--playlist` flag to remove tracks from an XSPF playlist file
+  instead of the current playlist.
+- `mpvd move -p`/`--playlist` flag to move tracks within an XSPF playlist file
+  instead of the current playlist.
 
 ## [2.5.0] - 2026-09-20